[#29] : feat : 줄거리 요약 구현
